Docstrings & Comments
Added docstrings and comments.

Fixed bug in get_action() method in secret_messages.py. Need to make
sure that action is returned after the method is re-run in the event of
an error. Otherwise, action stays stuck as bad value and causes program
to believe decrypt was chosen after an initial action entry error.